Use doubles instead of floats for more precise values

+++ OsuParsers/Beatmaps/Objects/Standard/StandardSlider.cs
@@ -10,12 +10,12 @@ namespace OsuParsers.Beatmaps.Objects.Standard
         public CurveType CurveType { get; set; }
         public List<Point> SliderPoints { get; set; } = new List<Point>();
         public int Repeats { get; set;  }
-        public float PixelLength { get; set; }
+        public double PixelLength { get; set; }
         public List<HitSoundType> EdgeHitSounds { get; set; }
         public Tuple<SampleSet, SampleSet>[] EdgeAdditions { get; set; }
 
         public StandardSlider(Point position, int startTime, int endTime, HitSoundType hitSound, bool isNewCombo, int comboOffset, CurveType type,
-            List<Point> points, int repeats, float pixelLength, List<HitSoundType> edgeHitSounds, Tuple<SampleSet, SampleSet>[] edgeAdditions, HitObjectExtras extras) 
+            List<Point> points, int repeats, double pixelLength, List<HitSoundType> edgeHitSounds, Tuple<SampleSet, SampleSet>[] edgeAdditions, HitObjectExtras extras) 
             : base(position, startTime, endTime, hitSound, isNewCombo, comboOffset, extras)
         {
             CurveType = type;
